+++ ErrorHandling/Result.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ResultOfTask
+namespace ResultOf
 {
     public class None
     {
@@ -8,6 +8,7 @@ namespace ResultOfTask
         {
         }
     }
+
     public struct Result<T>
     {
         public Result(string error, T value = default(T))
@@ -15,6 +16,11 @@ namespace ResultOfTask
             Error = error;
             Value = value;
         }
+        public static implicit operator Result<T>(T v)
+        {
+            return Result.Ok(v);
+        }
+
         public string Error { get; }
         internal T Value { get; }
         public T GetValueOrThrow()
@@ -36,6 +42,10 @@ namespace ResultOfTask
         {
             return new Result<T>(null, value);
         }
+        public static Result<None> Ok()
+        {
+            return Ok<None>(null);
+        }
 
         public static Result<T> Fail<T>(string e)
         {
@@ -54,25 +64,63 @@ namespace ResultOfTask
             }
         }
 
+        public static Result<None> OfAction(Action f, string error = null)
+        {
+            try
+            {
+                f();
+                return Ok();
+            }
+            catch (Exception e)
+            {
+                return Fail<None>(error ?? e.Message);
+            }
+        }
+
         public static Result<TOutput> Then<TInput, TOutput>(
             this Result<TInput> input,
             Func<TInput, TOutput> continuation)
         {
-            throw new NotImplementedException();
+            return input.Then(inp => Of(() => continuation(inp)));
+        }
+
+        public static Result<None> Then<TInput>(
+            this Result<TInput> input,
+            Action<TInput> continuation)
+        {
+            return input.Then(inp => OfAction(() => continuation(inp)));
         }
 
         public static Result<TOutput> Then<TInput, TOutput>(
             this Result<TInput> input,
             Func<TInput, Result<TOutput>> continuation)
         {
-            throw new NotImplementedException();
+            return input.IsSuccess
+                ? continuation(input.Value)
+                : Fail<TOutput>(input.Error);
         }
 
         public static Result<TInput> OnFail<TInput>(
             this Result<TInput> input,
             Action<string> handleError)
         {
-            throw new NotImplementedException();
+            if (!input.IsSuccess) handleError(input.Error);
+            return input;
+        }
+
+        public static Result<TInput> ReplaceError<TInput>(
+            this Result<TInput> input,
+            Func<string, string> replaceError)
+        {
+            if (input.IsSuccess) return input;
+            return Fail<TInput>(replaceError(input.Error));
+        }
+
+        public static Result<TInput> RefineError<TInput>(
+            this Result<TInput> input,
+            string errorMessage)
+        {
+            return input.ReplaceError(err => errorMessage + ". " + err);
         }
     }
 }
