--- ErrorHandling/Solved/ResultQueryExpressionExtensions.cs
@@ -1,23 +1,0 @@
-﻿using System;
-
-namespace ResultOf
-{
-    public static class ResultQueryExpressionExtensions
-    {
-        public static Result<TOutput> SelectMany<TInput, TOutput>(
-            this Result<TInput> input,
-            Func<TInput, Result<TOutput>> continuation)
-        {
-            return input.Then(continuation);
-        }
-
-        public static Result<TSelected> SelectMany<TInput, TOutput, TSelected>(
-            this Result<TInput> input,
-            Func<TInput, Result<TOutput>> continuation,
-            Func<TInput, TOutput, TSelected> resultSelector)
-        {
-            return input.Then(continuation)
-                .Then(o => resultSelector(input.Value, o));
-        }
-    }
-}
